+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/SpacingAroundUnaryOperatorRule.kt
@@ -12,8 +12,7 @@ import org.jetbrains.kotlin.com.intellij.lang.ASTNode
  *
  * @see [Kotlin Style Guide](https://kotlinlang.org/docs/reference/coding-conventions.html#horizontal-whitespace)
  */
-class SpacingAroundUnaryOperatorRule : Rule("unary-op-spacing") {
-
+class SpacingAroundUnaryOperatorRule : Rule("$experimentalRulesetId:unary-op-spacing") {
     override fun visit(
         node: ASTNode,
         autoCorrect: Boolean,